Organize query-related modules.
