+++ lib/orbit-common/cache.js
@@ -2,38 +2,36 @@ import Document from 'orbit/document';
 import Operation from 'orbit/operation';
 import Evented from 'orbit/evented';
 import { Class, clone, expose, isArray, isObject, isNone } from 'orbit/lib/objects';
-import { OperationNotAllowed } from './lib/exceptions';
 import { eq } from 'orbit/lib/eq';
-import CacheIntegrityProcessor from 'orbit-common/operation-processors/cache-integrity-processor';
-import SchemaConsistencyProcessor from 'orbit-common/operation-processors/schema-consistency-processor';
+import { OperationNotAllowed } from './lib/exceptions';
+import CacheIntegrityProcessor from './operation-processors/cache-integrity-processor';
+import SchemaConsistencyProcessor from './operation-processors/schema-consistency-processor';
 import Transform from 'orbit/transform';
 import TransformLog from 'orbit/transform-log';
 import { diffs } from 'orbit/lib/diffs';
-import QueryProcessor from 'orbit/query-processor';
+import QueryEvaluator from 'orbit/query/evaluator';
+import QueryProcessor from 'orbit/query/processor';
 import { normalizeOperations } from 'orbit/lib/operations';
 import {
   coalesceOperations,
   operationType,
   addRecordToSetOperation,
   removeRecordFromSetOperation
-} from 'orbit-common/lib/operations';
-import {
-  QueryEvaluator
-} from './oql/evaluator';
+} from './lib/operations';
 import {
   andOperator,
   orOperator,
   equalOperator,
   getOperator,
-  filterOperator,
+  filterOperator
+} from 'orbit/query/operators';
+import {
   relatedRecordsOperator,
   relatedRecordOperator
-} from './oql/operators';
-import { toIdentifier } from 'orbit-common/lib/identifiers';
-
-import LiveQueryOperators from './oql/live-query-operators';
+} from './query/operators';
+import { toIdentifier } from './lib/identifiers';
+import LiveQueryOperators from './query/live-query-operators';
 import { fromOrbitEvent } from './observable/adapters';
-
 
 /**
  `Cache` provides a thin wrapper over an internally maintained instance of a
