+++ lib/orbit-common/jsonapi-source.js
@@ -13,7 +13,7 @@ import CacheIntegrityProcessor from 'orbit-common/operation-processors/cache-int
 import DeletionTrackingProcessor from 'orbit-common/operation-processors/deletion-tracking-processor';
 import SchemaConsistencyProcessor from 'orbit-common/operation-processors/schema-consistency-processor';
 import { toIdentifier, parseIdentifier } from 'orbit-common/lib/identifiers';
-import QueryProcessor from 'orbit/query-processor';
+import QueryProcessor from 'orbit/query/processor';
 
 /**
  Source for accessing a JSON API compliant RESTful API with AJAX
