+++ lib/orbit-common/query/builder.js
@@ -1,16 +1,8 @@
-import { Class } from 'orbit/lib/objects';
-import { queryExpression as oqe } from 'orbit-common/oql/expressions';
-import OrbitQueryBuilder from 'orbit/query-builder';
-import {
-  TermBase,
-  Cursor
-} from 'orbit/query-builder/terms';
-import {
-  RecordCursor,
-  Records
-} from 'orbit-common/query-builder/terms';
+import { queryExpression as oqe } from 'orbit/query/expression';
+import OrbitQueryBuilder from 'orbit/query/builder';
+import { Records } from 'orbit-common/query/terms';
 
-class QueryBuilder extends OrbitQueryBuilder {
+export default class QueryBuilder extends OrbitQueryBuilder {
   constructor(options = {}) {
     super(options);
     this._typeTerms = this._createTypeTerms(options.terms.recordsOfType);
@@ -35,5 +27,3 @@ class QueryBuilder extends OrbitQueryBuilder {
     return typeTerms;
   }
 }
-
-export default QueryBuilder;
