+++ lib/orbit-common/query/live-query-operators.js
@@ -5,9 +5,7 @@ import {
 } from 'orbit-common/lib/operations';
 import { eq } from 'orbit/lib/eq';
 import { parseIdentifier } from 'orbit-common/lib/identifiers';
-import {
-  queryExpression as oqe
-} from 'orbit-common/oql/expressions';
+import { queryExpression as oqe } from 'orbit/query/expression';
 import { toIdentifier } from 'orbit-common/lib/identifiers';
 
 function filterByOql(operations, context, cache, oqlExpression) {
@@ -61,8 +59,7 @@ function filterByOql(operations, context, cache, oqlExpression) {
   });
 }
 
-
-var recordsOfType = {
+const recordsOfType = {
   evaluate(context, [type]) {
     return context.evaluator.target.patches.filter(operation => {
       return operation.path[0] === type;
@@ -70,7 +67,7 @@ var recordsOfType = {
   }
 };
 
-var filter = {
+const filter = {
   evaluate(context, [operationsExpression, filterExpression]) {
     const cache = context.evaluator.target;
     const operations = context.evaluate(operationsExpression);
@@ -79,7 +76,7 @@ var filter = {
   }
 };
 
-var record = {
+const record = {
   evaluate(context, [type, recordId]) {
     return context.evaluator.target.patches.filter(operation => {
       return operation.path[0] === type && operation.path[1] === recordId;
@@ -94,7 +91,7 @@ function extractRecordFromHasOne(value) {
   return value[id];
 }
 
-var relatedRecord = {
+const relatedRecord = {
   evaluate(context, [type, recordId, relationship]) {
     const patches = context.evaluator.target.patches;
 
@@ -135,7 +132,7 @@ var relatedRecord = {
   }
 };
 
-var relatedRecords = {
+const relatedRecords = {
   evaluate(context, [type, recordId, relationship]) {
     const patches = context.evaluator.target.patches;
     const initialMembersMap = context.evaluate(oqe('query:relatedRecords', type, recordId, relationship));
