+++ lib/orbit-common/query/terms.js
@@ -1,9 +1,8 @@
-import { queryExpression as oqe } from 'orbit-common/oql/expressions';
-
+import { queryExpression as oqe } from 'orbit/query/expression';
 import {
   Cursor,
   TermBase
-} from 'orbit/query-builder/terms';
+} from 'orbit/query/terms';
 
 class RecordCursor extends Cursor {
   attribute(name) {
