+++ lib/orbit-common/store.js
@@ -1,4 +1,7 @@
 import {
+  queryExpression as oqe
+} from 'orbit/query/expression';
+import {
   addRecordOperation,
   replaceRecordOperation,
   removeRecordOperation,
@@ -9,9 +12,6 @@ import {
   replaceHasManyOperation,
   replaceKeyOperation
 } from 'orbit-common/lib/operations';
-import {
-  queryExpression as oqe
-} from 'orbit-common/oql/expressions';
 import MemorySource from 'orbit-common/memory-source';
 
 export default MemorySource.extend({
