+++ lib/orbit/query/builder.js
@@ -1,7 +1,7 @@
-import { queryExpression as oqe } from 'orbit-common/oql/expressions';
-import { Value } from 'orbit/query-builder/terms';
+import { queryExpression as oqe } from './expression';
+import { Value } from './terms';
 
-class QueryBuilder {
+export default class QueryBuilder {
   constructor(options = {}) {
     this._options = options;
 
@@ -12,7 +12,6 @@ class QueryBuilder {
     return new Value(oqe('get', path));
   }
 
-  // should probably go somewhere else
   or(a, b) {
     return oqe('or', a, b);
   }
@@ -21,5 +20,3 @@ class QueryBuilder {
     return oqe('and', a, b);
   }
 }
-
-export default QueryBuilder;
