+++ lib/orbit/query/expression.js
@@ -1,6 +1,6 @@
 import { Class, isObject } from 'orbit/lib/objects';
 
-var QueryExpression = Class.extend({
+const QueryExpression = Class.extend({
   __oqe__: true,
   op: null,
   args: null,
@@ -15,16 +15,12 @@ var QueryExpression = Class.extend({
   }
 });
 
-function queryExpression(op) {
+export function queryExpression(op) {
   return new QueryExpression(op, Array.prototype.slice.call(arguments, 1));
 }
 
-function isQueryExpression(obj) {
+export function isQueryExpression(obj) {
   return isObject(obj) && obj.__oqe__;
 }
 
-export {
-  QueryExpression,
-  queryExpression,
-  isQueryExpression
-};
+export default QueryExpression;
