+++ lib/orbit/query/operators.js
@@ -1,11 +1,11 @@
 import {
   queryExpression as oqe
-} from './expressions';
+} from './expression';
 import {
   splitPath
-} from 'orbit/lib/paths';
+} from '../lib/paths';
 
-var andOperator = {
+const andOperator = {
   evaluate(context, args) {
     for (let arg of args) {
       if (!context.evaluate(arg)) {
@@ -16,7 +16,7 @@ var andOperator = {
   }
 };
 
-var orOperator = {
+const orOperator = {
   evaluate(context, args) {
     for (let arg of args) {
       if (!!context.evaluate(arg)) { return true; }
@@ -25,7 +25,7 @@ var orOperator = {
   }
 };
 
-var equalOperator = {
+const equalOperator = {
   evaluate(context, args) {
     let value;
     let valueSet = false;
@@ -43,7 +43,7 @@ var equalOperator = {
   }
 };
 
-var getOperator = {
+const getOperator = {
   evaluate(context, args) {
     let path = splitPath(args[0]);
     if (context.basePath) {
@@ -53,7 +53,7 @@ var getOperator = {
   }
 };
 
-var filterOperator = {
+const filterOperator = {
   evaluate(context, args) {
     const path = splitPath(args[0]);
     const where = args[1];
@@ -72,40 +72,10 @@ var filterOperator = {
   }
 };
 
-var relatedRecordsOperator = {
-  evaluate(context, [type, recordId, relationship]) {
-    const cache = context.evaluator.target;
-    const data = cache.get([type, recordId, 'relationships', relationship, 'data']);
-
-    const results = {};
-
-    Object.keys(data||{}).forEach(identifier => {
-      const [type, id] = identifier.split(':');
-      results[id] = cache.get([type, id]);
-    });
-
-    return results;
-  }
-};
-
-var relatedRecordOperator = {
-  evaluate(context, [type, recordId, relationship]) {
-    const cache = context.evaluator.target;
-    const data = cache.get([type, recordId, 'relationships', relationship, 'data']);
-
-    if (!data) { return null; }
-
-    const [relatedType, relatedRecordId] = data.split(':');
-    return { [relatedRecordId]: cache.get([relatedType, relatedRecordId]) };
-  }
-};
-
 export {
   andOperator,
   orOperator,
   equalOperator,
   getOperator,
-  filterOperator,
-  relatedRecordsOperator,
-  relatedRecordOperator
+  filterOperator
 };
