+++ lib/orbit/query/processor.js
@@ -1,4 +1,4 @@
-import { QueryProcessorNotFoundException } from './lib/exceptions';
+import { QueryProcessorNotFoundException } from '../lib/exceptions';
 
 export default {
   init() {
